+++ DossierC/gestion.h
@@ -11,7 +11,7 @@ void afficherShiftsReduits(Shift * pDebShift);
 Boolean rechercheShift(Shift * pDebShift, int choixDate, int choixHeure, Shift * * pPShift);
 int rechercheMembreDansTableau(int num,  Membre * tabMembres);
 Boolean verifDoubletteExisteDansShift(Inscription * pDebListeIns, int num1, int num2);
-Erreur ajoutInscription(Shift * * pShift, int num1, int num2, double categorie);
+Erreur ajoutInscription(Shift * * pShift, int num1, int num2, int categorie);
 void afficherDoublettesShift(Shift * pShift);
 Boolean rechercheDoubletteDansShift(Inscription * pDebListeIns, Inscription * * pDoublette, Inscription * * pDoubletteSauve, int id);
 Erreur suppressionInscription(Shift * * pShift, Inscription * * pDoublette, Inscription * * pDoubletteSauve);
